fix: use `${{ runner.temp }}` instead of `$HOME`
We change the location to store garden-action temporary files to use ${{ runner.temp }}. This directory is being cleaned before and after the job execution. 

This is helpful in case this action is being used on self-hosted runners, and stops polluting the home directory of self hosted runners in case the job is not being executed in isolated VMs.